% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanesTablesInfo.R
\name{nhanesTablesInfo}
\alias{nhanesTablesInfo}
\title{Show a basic information of NHANCE table}
\usage{
nhanesTablesInfo(x, detail=FALSE, showfactor=3, noMedian = F, Hid.val1="", Hid.val2="")
}
\arguments{
\item{x}{The name of the specific table to retrieve}

\item{detail}{Option to show proportion for factors and median(min, max) for integers}

\item{showfactor}{Maximum levels to show for one variable}

\item{Hid.val1}{Option to hid some values}

\item{Hid.val2}{Option to hid some values}

\item{noMedian}{Option to hid or show median(min, max) in detailed type table}
}
\value{
The table is returned as a data frame with list.
}
\description{
This function shows all the basic information for NHANSE table.
}
\details{
This function shows all the basic information for NHANSE table. There are two types of
 choices to show information for a table. For the first type, this table shows the number of variables
 in the table, the meaning and types of the variables. In the Value column in the summary table, the
 SEQN will always show the range of respondent sequence number. For the categorical variables, the
 Value column only takes smaller than three levels variables which have less or equal to 25 characters
 in total, otherwise the Value column will show the number of different levels for one categorical
 variable. For the continuous variables, the Value column will show the mean and standard deviation.
 Last but not least, both types of table will show missing value and valid values. For the second type,
 it will not show the meaning of the variables in the table, the meaning of the variables will go to
 another list. However, the second type of tables can show more information like the proportion and
 median, minimum and maximum values.
}
\examples{
\dontrun{
nhanesTablesInfo(table, detail=F)
nhanesTablesInfo(table, detail=T, showfactor=15)
nhanesTablesInfo(table, detail=T, noMedian = T, Hid.val1="Yes", Hid.val2="No")
}
}
\keyword{summary}
