% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationGraph.R
\name{relationGraph}
\alias{relationGraph}
\title{Visualization of association between variables}
\usage{
relationGraph(dataset, alpha = alpha, table = T, graph = T, show.p = F, vertex.size = 20, 
                     vertex.label.cex=1.3, vertex.label.font = 2, vertex.shape = "circle", 
                     edge.curved = 0, edge.width = 2)
}
\arguments{
\item{dataset}{Dataset going to be analyzed}

\item{alpha}{User-defined significance level}

\item{table}{Display relation table or not}

\item{graph}{Display relation graph or not}

\item{show.p}{Display p-values in the relation graph or not}

\item{vertex.size}{Vertex size}

\item{vertex.label.cex}{Vertex label size}

\item{vertex.label.font}{Vertex label font style (e.g. 'bold')}

\item{vertex.shape}{Vertex shape (e.g. 'circle', 'rectangle')}

\item{edge.curved}{Curvature of edge (1 = straight line)}

\item{edge.width}{Width of edges}
}
\description{
Display the relationships or associations between each pairs of the variables.
}
\details{
Table and Graph of Relationships/Associations

The function provides a idea of the associations between each pairs of the variables in the given dataset. 
         A table of pairs that associations are significant at a user-defined statistical significance level and a 
         corresponding relation graph will be displayed. Significance level is user-defined. Only the associations 
         that has a p-value less than alpha will be displayed. Note that the graph might not be clear if there exists 
         too many associations. Dashed edges denote that the p-value of the comparisons are greater than 0.05 while 
         solid edges are less than 0.05.
}
