% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSize.R
\name{sampleSize}
\alias{sampleSize}
\title{Sample Size Table}
\usage{
sampleSize(E.start=h1, power.start=pr1, E.end=NULL, power.end=NULL, type="prop",
                   sig.level=0.05, alternative = "two.sided")
}
\arguments{
\item{E.start}{Minimum effect size}

\item{power.start}{Minimum power of test}

\item{E.end}{Maximum effect size}

\item{power.end}{Maximum power of test}

\item{type}{Type of test : proportion test or t.test ("prop" / "t.test")}

\item{sig.level}{Significance level (Type I error probability)}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less"}
}
\value{
sample size table
}
\description{
This function gets a range of effect size and power from user and return a
10 x 10 matrix of recommended sample size.
}
\examples{
\dontrun{
sampleSize(E.start=0.1, power.start=0.8)
sampleSize(E.start=0.1, E.end=0.3, power.start=0.8, power.end=0.95, sig.level=0.05)
}
}
\keyword{sample}
\keyword{size}
