% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showInfo.R
\name{showInfo}
\alias{showInfo}
\title{Get a merged NHANCE table with detail}
\usage{
showInfo(...)
}
\arguments{
\item{tablenames}{The name of the specific table to retrieve. (Multiple tables are allowed)}
}
\value{
The table is returned as a data frame.
}
\description{
This function download table from NHANCE website with character formate like YES/NO and Female/Male
instead of 0, 1. Moreover, it can download and merge mutile tables at the same time. But users need to give the
right table names(which can get by function in nhanesA and NHANES offical website)
}
\examples{
\dontrun{
showInfo(DEMO_C)
showInfo(L11PSA_C, DEMO_C, PSA_D, DEMO_D)
}
}
\keyword{NHANES}
