% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableTwo.R
\name{tableTwo}
\alias{tableTwo}
\title{Additional relationship check between variables}
\usage{
tableTwo(formula, table, Multi=TRUE, CI=TRUE, poisson=FALSE)
}
\arguments{
\item{formula}{an object of class "formula" (y ~ x1 + x2 + x3):
a symbolic description of the model to be fitted.}

\item{table}{a dataframe generated by showInfo function in this package.}

\item{Multi}{option to apply univariate analysis or multivariable analysis.}

\item{CI}{option to show confidence interval in the table.}

\item{poisson}{option to apply poisson regression.}
}
\value{
a univariate or multivariable data frame
}
\description{
This function applies different regression model based the type of response variable
to do additional relationship check.
}
\details{
This function test applies a regression model based the response variable.
If the response variable in the formula is categorial with two levels, then it will apply
binomial logit function. User can choose to apply multivariate or univariate test. As the
result, this function will return a table and a odds plot. If the response variable has
more than 2 levels, then this function will stop and tell user, we cannot deal with it.
When the function detects the response variable is continuous, it will apply Wilcox test,
T-Test, ANOVA test and linear regression. And user also can choose to apply poisson
regression by parameter poisson.
}
\examples{
\dontrun{
tableTwo(formula = INDFMPIR ~RIAGENDR+ DMDBORN + RIDAGEMN, table=tableName, Multi=T)
tableTwo(formula = RIAGENDR ~ INDFMPIR + DMDBORN + RIDAGEMN, table=tableName, Multi=F)
}
}
\keyword{Check}
\keyword{Realtionship}
